refactor: extract approval workflow from tool_executor (Phase 10)
ApprovalWorkflow class — pattern learning, HITL prompts, safety gates,
MCP/write/cost/bash confirmation. Composition via ToolExecutor._approval.

tool_executor.py 1390→862 lines (-38%)

Quality: ruff 0, mypy 0 (211 files), pytest 3871 pass
